+++ Cargo.toml
@@ -14,6 +14,8 @@ edition = "2021"
 [dependencies]
 bytemuck = "1.13.1"
 cgmath = "0.18.0"
+egui-wgpu = "0.23.0"
+egui-winit = "0.23.0"
 env_logger = "0.10.0"
 futures-intrusive = "0.5.0"
 noise = "0.8.2"
