+++ src/rendering/gui.rs
@@ -1,0 +1,14 @@
+pub struct GuiRenderer
+{
+
+}
+
+pub struct GuiRenderStage
+{
+
+}
+
+pub struct GuiDrawCall
+{
+    
+}
